Build Windows support directly in UciShell
Previously, Windows support for remote engines required patching the
spur library. This made it difficult to update the dependency.

This change includes the WindowsShellType to pass to spur, with some
modifications to support the kill command. Also, UciShell is now acting
as a proxy to the underlying shell. This lets UciShell act as the
underlying SshShell, and will enable further refactoring.

Finally, the __init__ method of UciShell was simplified to remove some
duplication.