Do not create a PicoTutor instance on State initialization
Before this change, initializing the State created a default PicoTutor
instance, spawning two `a-stockf` processes that went unused.

This change will prevents this from happening.

There are still problems with the implementation where a PicoTutor will
still be instantiated with 2 engines even if not using the feature, that
will require further changes.