Undo last fix (partially)
Unfortunately there is still a situation where the LEDs flash in a wrong(mirrored) direction. This seems to be a hardware issue of the REV2 DGT eBoard and not picochess. Workaround: if you want to play white with black pieces in front of you (what seems kind of odd ;-) then FIRST play with black pieces one move restart new game and from then on it works correctly . This has nothing to do with picochess (you can reboot picochess and still still works until you switch off the Rev2).