Reset of original time settings after switching back to an engine without pico uci time parameter
When using the new PicoChess UCI parameter PicoNode, PicoDepth or PicoTimeControl for an engine the clock is set to the corresponding time settings automatically. After switching back to an engine without these uci pico time parameters the time settings are set to the original time setting before this automatic change.